fix: auth logo, sqlite3 library loading, and sync color_argb overflow
- Auth screens: replace blank icon with Image.asset app logo (login + signup)
- Copy mipmap-xxxhdpi/ic_launcher.png to assets/images/app_icon.png
- Android: add packaging.jniLibs.useLegacyPackaging=true to extract .so files
  at install time, fixing libsqlite3.so not found error
- Drift: downgrade sqlite3_flutter_libs 0.6.0+eol -> 0.5.41 and call
  applyWorkaroundToOpenSqlite3OnOldAndroidVersions() on Android
- Backend schema: change tasks.color_argb INT -> BIGINT (Flutter ARGB values
  are unsigned 32-bit, overflowing signed INT -> 500 on task sync)
- Backend zod: remove .int() constraint on color_argb to allow BIGINT range
- Backend migrate.js: add idempotent ALTER TABLE migration for live Render DB
- Sync: mask color_argb with 0xFFFFFFFF before sending to ensure valid range

+++ backend/db/schema.sql
@@ -73,7 +73,7 @@ CREATE TABLE IF NOT EXISTS tasks (
   description                 TEXT,
   purpose                     TEXT,
   icon_id                     TEXT,
-  color_argb                  INT,
+  color_argb                  BIGINT,
   tags_json                   TEXT        NOT NULL DEFAULT '[]',
   start_minutes               INT,
   duration_minutes            INT,
